+++ test/IntegrationTests/LibraryVersions.g.cs
@@ -267,6 +267,19 @@ public static partial class LibraryVersion
             return theoryData;
         }
     }
+    public static TheoryData<string> RabbitMq
+    {
+        get
+        {
+            var theoryData = new TheoryData<string>();
+#if DEFAULT_TEST_PACKAGE_VERSIONS
+            theoryData.Add(string.Empty);
+#else
+            theoryData.Add("7.0.0-rc.1");
+#endif
+            return theoryData;
+        }
+    }
     public static TheoryData<string> StackExchangeRedis
     {
         get
@@ -345,6 +358,7 @@ public static partial class LibraryVersion
        { "OracleMda", OracleMda },
        { "OracleMdaCore", OracleMdaCore },
        { "Quartz", Quartz },
+       { "RabbitMq", RabbitMq },
        { "StackExchangeRedis", StackExchangeRedis },
        { "WCFCoreClient", WCFCoreClient },
        { "Kafka", Kafka },